// src/lib/api/base/api.ts
export interface CalendarEventEnum {
}
// ...
export interface CustomCalendarData {
    /**
     * 
     * @type {any}
     * @memberof CustomCalendarData
     */
    'start': any;
    /**
     * 
     * @type {any}
     * @memberof CustomCalendarData
     */
    'end': any;
    /**
     * 
     * @type {any}
     * @memberof CustomCalendarData
     */
    'allDay': any;
    /**
     * 
     * @type {any}
     * @memberof CustomCalendarData
     */
    'title': any;
}
// ...
export interface DaysEnum {
}
/**
 * 
 * @export
 * @interface EventSchemaInput
 */
export interface EventSchemaInput {
    /**
     * 
     * @type {any}
     * @memberof EventSchemaInput
     */
    'data': any;
    /**
     * 
     * @type {any}
     * @memberof EventSchemaInput
     */
    'event_date': any;
}
/**
 * 
 * @export
 * @interface EventSchemaOutput
 */
export interface EventSchemaOutput {
    /**
     * 
     * @type {any}
     * @memberof EventSchemaOutput
     */
    'name': any;
    /**
     * 
     * @type {any}
     * @memberof EventSchemaOutput
     */
    'data': any;
    /**
     * 
     * @type {any}
     * @memberof EventSchemaOutput
     */
    'event_date': any;
}
// ...
export interface SegmentEnum {
}
// ...
export class CalendarApi extends BaseAPI {
    /**
     * Super Admin only can add event
     * @summary Add Event
     * @param {CalendarEventEnum} eventType 
     * @param {EventSchemaInput} eventSchemaInput 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CalendarApi
     */
    public addEventV1CalendarEventsPost(eventType: CalendarEventEnum, eventSchemaInput: EventSchemaInput, options?: AxiosRequestConfig) {
        return CalendarApiFp(this.configuration).addEventV1CalendarEventsPost(eventType, eventSchemaInput, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Daily Events
     * @param {any} startDate 
     * @param {any} endDate 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CalendarApi
     */
    public dailyEventsV1CalendarDailyEventsGet(startDate: any, endDate: any, options?: AxiosRequestConfig) {
        return CalendarApiFp(this.configuration).dailyEventsV1CalendarDailyEventsGet(startDate, endDate, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Daily Ohlc
     * @param {any} symbol 
     * @param {SegmentEnum} segment 
     * @param {any} startDate 
     * @param {any} endDate 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CalendarApi
     */
    public dailyOhlcV1CalendarDailyOhlcGet(symbol: any, segment: SegmentEnum, startDate: any, endDate: any, options?: AxiosRequestConfig) {
        return CalendarApiFp(this.configuration).dailyOhlcV1CalendarDailyOhlcGet(symbol, segment, startDate, endDate, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get Custom Ohlc Data
     * @param {any} indexName 
     * @param {any} resultType 
     * @param {any} fromDate 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CalendarApi
     */
    public getCustomOhlcDataV1CalendarCustomOhlcDataGet(indexName: any, resultType: any, fromDate: any, options?: AxiosRequestConfig) {
        return CalendarApiFp(this.configuration).getCustomOhlcDataV1CalendarCustomOhlcDataGet(indexName, resultType, fromDate, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary List Events
     * @param {DaysEnum} daysBefore 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CalendarApi
     */
    public listEventsV1CalendarEventsGet(daysBefore: DaysEnum, options?: AxiosRequestConfig) {
        return CalendarApiFp(this.configuration).listEventsV1CalendarEventsGet(daysBefore, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Ohlc
     * @param {any} symbol 
     * @param {SegmentEnum} segment 
     * @param {DaysEnum} daysBefore 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CalendarApi
     */
    public ohlcV1CalendarOhlcGet(symbol: any, segment: SegmentEnum, daysBefore: DaysEnum, options?: AxiosRequestConfig) {
        return CalendarApiFp(this.configuration).ohlcV1CalendarOhlcGet(symbol, segment, daysBefore, options).then((request) => request(this.axios, this.basePath));
    }
}